fix: the lzma decoder seems to always fail with a memory error
I was unable to use the lzma decompressor even though 7zip was able to identify them as lzma compressed and decompress them without a problem. Under the debugger the lzma_code() call was returning `LZMA_MEMLIMIT_ERROR`.

I found online that the error can be fixed using `lzma_memlimit_set()` using a value obtained with `lzma_memusage()`. I tried to avoid having to call `lzma_code()` twice but if the functions are called before the first `lzma_code()`the values of memory obtained fall short and error occurs again.

I suspect that there are better ways to deal with this other than the code proposed in this PR, but I haven't been able to find any.