fix: Avoid crashing imhex when passing invalid data to the 3-d visualizer.
There are some obvious constrains in the various array data sizes that we can use to determine if their sizes and in some cases their values are correct. By far, the most important of the various arrays are the indices, so their definition and use has been simplified a great deal. In order to treat all variables in a similar manner the changes to the uv variable posted in an earlier PR were also added here.

With this one, all the recently opened PR's combined should contain exactly the same changes and fixes as the older 3-d visualizer PR 1850 which I will close shortly after this one is opened.