Added underwaved text functions.
Switched error markers from highlighting entire line to underwaving the error location. Definition or error marker was switched to include column and size as well as line and message like before. This change required changing the way view pattern editor draws the error markers because the errors themselves don't have size information. Also, a new errorHoverBoxes type was defined to help in the detection of the floating error messages when error is hovered.

Note that the underwave code depends on having a monospaced. If font is not monospaced the underwaved text can be short/long or displaced.