fix: Couldn't check for window focus from within a shortcut callback.
this was caused by the call to the shortcut manager callback processing occurring outside an ImGui::Begin() ImGui::End() block for the view in question.
improv: Allow the find/replace class to work on any TextEditor instance defined. Now the TextEditor can be passed as an argument to the entire find/replace subsystem with come configuration options like the ability to turn replace off completely. This can be useful for other texteditor instances like http requests and view bookmarks
fix: It is not possible to copy the text from the console. This was caused by the move of the keyboard handling from the text editor to view_pattern_editor. The previous bug was preventing a proper fix. Once that error was fixed it was possible to test for focus in sub-windows of the views so that shortcuts could act differently depending on where focus was located. For context menus we create sub window hover boxes to decide which context menu to draw.
improv: Context menus and shortcuts were added to the console and the virtual filesystem sub windows of the pattern editor view. Additions include a search popup for the console and an open as view context menu entry for files in the virtual filesystem.
fix: Files that had the same path in the virtual filesystem were sharing the same ImGui ID which generated menu entries for all of them in the context menu.