fix: finding utf-8 strings failed to select their location.
The code was using the byte index of the match into the utf-8 string to store the match locations, but the code that sets the selection uses the char index into the utf-8 string instead. Another problem was that the search uses the byte index to determine if it needs to find more matches. Both problems were solved by introducing two functions to switch from coordinates in units of bytes to coordinates in units of chars and vice versa.

fix: The previous fix also solved the (unreported) bug of being unable to select utf-8 works by double-clicking.

fix: The move to next/previous word (Ctrl-arrow) behaved differently depending on the direction. I made both the move left/right functions share a much simpler algorithm and rewrote the find start/end of word functions share the same codestructure.