fix: Reset cursor blink on move
When this feature was implemented the text editor handled all the keyboard input so the blink reset function could be called in one place for all the key presses. Now that shortcuts are done in the shortcut manager we need to reset the blink state inside all the functions that can affect cursor movement.  There may be more functions that need this that havent been merged but git should show merging conflicts so it should be easy to deal with.