fix: Missing runtime error markers
Moving error markers to be underwaved created s couple of errors:
Markers that were 1 char long and located at the end of the line didn't show on text editor.
Markers that had zero length or no stack trace didn't show on text editor.

The first error was caused by an off by one error in the column index.
There was no implementation for errors that had no location or zero length. Now errors with no location will be shown at the beginning of the line where they occur. Errors with zero length will be marked from their location to the end of the line, therefore errors with no stack trace will be marked from the beginning to the end of the line.