fix: breakpoints become disabled after first hit.
If you set a breakpoint the first time you run the pattern it gets hit. After that subsequent runs will fail to hit the breakpoint.

fix: pressing F5 (start pattern execution ) while being in a breakpoint hangs ImHex.

The first bug was caused by two factors.
- When breakpoint hits evaluation stops before the line with the breakpoint is evaluated. In order to avoid hitting the breakpoint again when evaluation is continues the breakpoint needs to be disabled. Once the line is evaluated the breakpoint needs to be enabled again, but the previous code didn't do that.
- Detecting if the line was being evaluated used insufficient information to perform said detection. Because of the way AST nodes are assigned coordinates, Nodes that lie in other lines iof source code may and up sharing the same line number as the Nodes in the breakpoint. To tell them apart one must also use columns

The reason for hanging ImHex was that the shortcut procedure was not checking if the pattern was already running, and it attempted to start another one. This makes ImHex wait indefinitely for a lock to be released

To fix the first issue both problems were resolved. When evaluation of another line occurs the breakpoint is re-enabled. Note that at this point the full evaluation of the line may not have finish yet. This is of no concern as the second problem prevents triggering the breakpoint erroneously. By using the column of the position of the node we ensure that only valid evaluations of the node trigger the breakpoints.

To fix the hanging of ImHex we check the runtime for current evaluation and if detected we stop it. In all cases the evaluation is started again.