fix: fixes for issue #1704.
Loading a file on the command line or using the context menu in windows was triggering an achievement before the achievements file was loaded. This resulted on the achievements.json file to be overwritten with all the achievements reset except for the one to open files.
This PR fixes the problem by introducing a boolean that is used to check if the file has been loaded.