Inline visualizers in custom data inspectors
This PR aims at making inline visualizers work on the data inspector so that more rgb encodings can be added as custom pattern language inspector rows. This was never setup to work because the inline visualizer rendering function was a private member function.

In order to be accessible from the inspector class the rendering function was made public.

Missing still is the tooltip to make it behave like other color entries.