A better implementation for the inline visualizers in the data inspector.
The previous version was trying to use the drawVisualizer function in PatternDrawer from DataInspector which was inefficient. The former drawVisualizer code was put in its own class in ui from imhex plugin library. Both DataInspector and PatternDrawer get a member field that is used to draw inline and regular visualizers. Code has been tested to make sure it works for both the inspector and the pattern drawer.