Sound visualizer window rendered incorrectly.
The sliding bar and the play controls were being drawn over each other.
Added milliseconds to the timer. The end of the sliding bar was off by one.
Changes to make code more efficient and better formatted.

To fix the problem I included the last two rows to the subplot block so that ImPlot can draw them each on its own line.