fix: empty documents have no cursor.
Even if focus is given to the pattern editor the cursor will not show until it contains printable chars. I was using the wrong condition to enter the loop that skipped the default empty document. The fix was to check if there are lines of text that need to be processed even if they are empty.