feat: added goto line popup to text editors.
Since the popup is fairly small I opted for a straight addition parallel to the find/replace. To make code more clear the functions that create each popup were coalesced and made their interface simpler. That forced a reorganization of the data processing which translates to a larger number of changes than usual. Most of those changes are just moving some action from one function to another. The old method to identify popups using the size and position of the window was dropped in favour of one based on child windows abd using their names for a much easier and robust identification.
Added specialized functions to texteditor to jump to a line or to given coordinates with a simple interface that simplifies older code that performed the same task.
Because this PR modifies heavily the same code as the previous PR (1983) it is also included here to make merging easier.