Feat: Added clickable links to error messages.
Errors printed in the console can be clicked to have the cursor jump to the source code line where the error is at. The mouse cursor changes its shape to indicate which parts can be clicked. When the cursor jumps, the text editor takes the focus away from the console, and it scrolls the window to make the line with the error visible if it isn't.
This code uses the function created for the goto PR but adds code to switch focus to the target