improv: VScode like scrolling
I have implemented a fix that makes the scrollbars behave like they do in vscode. Vertically you can scroll past the end of the file until only the last line can be seen at the top. The horizontal bar behaves as if every line was the same length which is the length of the longest line in the file. Not only this creates a better user experience, but it also fixes the annoying flicker that occurs when scrolling through large files.
Also, I have switched all the old draw calls to render text to regular TextUnformatted calls which adds extra stability to the resulting display.