Improved scrolling
Two major improvements:
1) see through scrollbars when not hovered.
2) un-scrollable line numbers.

Also enlarged display region by eliminating padding.
There is still a problem with lines jumping when the scrollbar is dragged but it is limited to one line and probably due to floating point error for scroll bar number. It is much less noticeable than the previous jumping which could involve several pages.