fix: find jumps to wrong location.
After successfully finding matches and setting the cursor to them, the screen would jump to the original window location upon closing the window.

The error was caused by the wrong assumption that the scroll location should be restored when window is closed. Instead, the right ammount of scrolling needs to be calculated to account for the window no longer covering part of the text editor. Unused variable was discarded.

Another unrelated error is that the history of search names cannot be accessed which will be addressed at a later PR.