fix: fixes for issue #2084
Error in text editor prevented using shortcuts when language other than English was chosen. The code was mistakingly using localized name to test which window had focus a execute the shortcut. Fixed it by switching the name of the child vindow to the constant value used to check the windows focused identity.