Align icons.
Icons become unaligned because the default font's descent is not calculated until the atlas is built. Atlas building must be delayed until custom rects are defined and besides, building an atlas just to obtain the descent is something that ImGui font implementation makes necessary.

All font and glyph metrics are accessible when they are loaded by specifying some desired size using the freetype api. It takes a little bit of extra code, but it is the most efficient way to obtain metrics.