fix: auto evaluate doesn't work with undo.
The pattern editor was resetting the flag that marked text changes on every frame, but resolving a text change may take more than one frame, so instead I created a function that allows the resetting of the text changed boolean from view pattern editor.

Currently, when a pattern is evaluated the pattern editor will lose focus regardless of how evaluation was triggered. This is specially annoying when using f5 or auto-evaluate. Now whenever a pattern is evaluated, focus is given to the pattern editor when evaluation ends.