fix: Console editor missed printed output.
This change is to fix a bug reported in discord by berkus and Naheulf about the console missing output lines. The bug was caused by using SetText which replaces the existing text with the text in the argument.

To fix it use InsertText which puts the text at the current cursor position that was already set to the end of the current contents.

Code was tested with pattern used to reproduce the bug and seemed to work when evaluated repeatedly.