Subpixel rendering for opengl backend of imgui using freetype implementation.
Most, if not all, the text tendering in imgui is done through RenderText so callbacks are inserted before and after all text related draw calls to turn on and off special settings that control subpixel rendering. Dual source  blending is required in this case and depending on the theme color we use the srgb render buffer or not. A parameter with Gamma correction is passed using an unused matrix element of the projection matrix (2d vs 3d) to the shaders which can use it to determine if subpixel rendering code should be executed. To compensate for the nonlinear color blending that has been the de facto standard a gamma correction is applied to the color of the text.

On the freetype side a bitmap with three times the sampling frequency in the x direction is first produced. the bitmap is further processed using the standard 5 tap lcd filter and the resulting glyphs are loaded into custom rects of the atlas.

The resulting rendered text is clearly sharper than plain anti-aliased and there are no color fringes that can be detected unless considerable zoom is used.

The pr will remain as a draft until further testing of various types of fonts and sizes can be performed, but it should compile and run fine under the master code base.