feat: made undo work when importing pattern.
A user complained that they imported a file by accident when they meant to export it and as a result had trouble recovering the changes they were trying to save. Autosave saved the day but there is no reason for not being able to undo changes after importing a pattern. Previously importing a pattern was treated as a reset on the editor instance which actually erased all previous undo entries. Importing now is treated as a normal editing where the entire file is replaced with the imported.